+++ src/main.rs
@@ -20,9 +20,28 @@
 // needs pub to generate documentation.
 pub mod request_api;
 
-fn main() -> Result<(), request_api::Error> {
-    let hexbot = request_api::fetch(3)?;
-    println!("Hexbot responded with this colors {}.", hexbot);
+fn main() {
+    let hexbot = request_api::fetch(5).unwrap();
+    println!("A hexbot with five colors: {}", hexbot);
 
-    Ok(())
+    let mut red_sum = 0;
+    for color in hexbot.colors() {
+        red_sum += color.to_rgb255().0 as i32;
+    }
+    println!("The sum of all red values: {}", red_sum);
+
+    let hexbot_with_coordinates = request_api::fetch_with_coordinates(5, 100, 100).unwrap();
+    println!(
+        "A hexbot with five colors and coordiantes: {}",
+        hexbot_with_coordinates
+    );
+
+    let coordinate = hexbot_with_coordinates.coordinates().unwrap()[1];
+    let color = hexbot_with_coordinates.colors()[1];
+    println!(
+        "The second color at position ({x}|{y}) has a blue component of {blue:.2}%.",
+        blue = color.blue,
+        x = coordinate.0,
+        y = coordinate.1
+    );
 }