+++ src/request_api.rs
@@ -21,6 +21,7 @@
 //!
 //! # Examples
 //!
+//! ## 1
 //! ```
 //! let hexbot: Hexbot = fetch(3).unwrap();
 //! println!("{}", hexbot); // [#RRGGBB, #RRGGBB, #RRGGBB]
@@ -37,6 +38,32 @@
 //! dbg!(color.to_hsl());    // (f64, f64, f64): (H, S, L)
 //! dbg!(color.to_yiq());    // (f64, f64, f64): (Y, I, Q)
 //! ```
+//! ## 2
+//! ```
+//! let hexbot = fetch(5).unwrap();
+//! println!("A hexbot with five colors: {}", hexbot);
+//!
+//! let mut red_sum = 0;
+//! for color in hexbot.colors() {
+//!     red_sum += color.to_rgb255().0 as i32;
+//! }
+//! println!("The sum of all red values: {}", red_sum);
+//!
+//! let hexbot_with_coordinates = fetch_with_coordinates(5, 100, 100).unwrap();
+//! println!(
+//!     "A hexbot with five colors and coordiantes: {}",
+//!     hexbot_with_coordinates
+//! );
+//!
+//! let coordinate = hexbot_with_coordinates.coordinates().unwrap()[1];
+//! let color = hexbot_with_coordinates.colors()[1];
+//! println!(
+//!     "The second color at position ({x}|{y}) has a blue component of {blue:.2}%.",
+//!     blue = color.blue,
+//!     x = coordinate.0,
+//!     y = coordinate.1
+//! );
+//! ```
 //!
 //! [Hexbot-API]: https://noopschallenge.com/challenges/hexbot
 
@@ -46,13 +73,20 @@ use std::fmt;
 use tint::Color;
 
 #[derive(Clone, Copy, Debug, Deserialize, PartialEq)]
+struct Coordinate {
+    x: i32,
+    y: i32,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq)]
 struct Dot {
     #[serde(rename = "value", deserialize_with = "deserialize_color")]
     color: Color,
+    coordinates: Option<Coordinate>,
 }
 
 /// Deserialized response
-#[derive(Debug, Deserialize, PartialEq)]
+#[derive(Clone, Debug, Deserialize, PartialEq)]
 pub struct Hexbot {
     colors: Vec<Dot>,
 }
@@ -62,17 +96,66 @@ impl Hexbot {
     pub fn colors(&self) -> Vec<&Color> {
         self.colors.iter().map(|dot| &dot.color).collect()
     }
+
+    /// Get the coordiantes of all colors.
+    ///
+    /// Returns `None` if the hexbot doesn't have coordinates,
+    /// otherwise it returns the coordinates in the form `(x, y)`
+    /// in an vector.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// let (x, y) = hexbot.coordinates().unwarp().get(1).unwrap();
+    /// println!("x: {}, y: {}", x, y);
+    /// ```
+    pub fn coordinates(&self) -> Option<Vec<(i32, i32)>> {
+        let mut vec = Vec::new();
+        for dot in &self.colors {
+            let Coordinate { x, y } = dot.coordinates?;
+            vec.push((x, y));
+        }
+        Some(vec)
+    }
+
+    /// Returns `true` if the hexbot has coordinates, otherwise `false`.
+    // #[inline]
+    pub fn has_coordinates(&self) -> bool {
+        !(self.colors[0].coordinates == None)
+    }
 }
 
 impl fmt::Display for Hexbot {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "[")?;
         let len = self.colors.len() - 1;
-        for (counter, Dot { color }) in self.colors.iter().enumerate() {
-            if counter == len {
-                write!(f, "{}]", color.to_hex().to_uppercase())?;
-            } else {
-                write!(f, "{}, ", color.to_hex().to_uppercase())?;
+        if self.has_coordinates() {
+            for (counter, Dot { color, coordinates }) in self.colors.iter().enumerate() {
+                if counter == len {
+                    write!(
+                        f,
+                        "{}-({}|{})]",
+                        color.to_hex().to_uppercase(),
+                        coordinates.unwrap().x,
+                        coordinates.unwrap().y
+                    )?;
+                } else {
+                    write!(
+                        f,
+                        "{}-({}|{}), ",
+                        color.to_hex().to_uppercase(),
+                        coordinates.unwrap().x,
+                        coordinates.unwrap().y
+                    )?;
+                }
+            }
+        } else {
+            for (counter, Dot { color, .. }) in self.colors.iter().enumerate() {
+                if counter == len {
+                    write!(f, "{}]", color.to_hex().to_uppercase())?;
+                } else {
+                    write!(f, "{}, ", color.to_hex().to_uppercase())?;
+                }
             }
         }
         Ok(())
@@ -87,12 +170,34 @@ fn deserialize_color<'d, D: Deserializer<'d>>(deser: D) -> Result<Color, D::Erro
 ///
 /// `count` must be between 1 and 1000.
 pub fn fetch(count: i32) -> Result<Hexbot, Error> {
-    if 0 < count && count <= 1000 {
+    if 1 <= count && count <= 1000 {
         Ok(reqwest::get(&format!(
             "https://api.noopschallenge.com/hexbot?count={}",
             count
         ))?
         .json()?)
+    } else {
+        Err(Error::CountOutOfRange)
+    }
+}
+
+/// Fetch the given count of colors in the given range of coordinates from the Hexbot API.
+///
+/// ## Thresholds:
+///  * `1 <= count <= 1,000`
+///  * `10 <= width <= 100,000`
+///  * `10 <= height <= 100,000`
+pub fn fetch_with_coordinates(count: i32, width: i32, height: i32) -> Result<Hexbot, Error> {
+    if 1 <= count && count <= 1000 {
+        if 10 <= width && width <= 100_000 && 10 <= height && height <= 100_000 {
+            Ok(reqwest::get(&format!(
+                "https://api.noopschallenge.com/hexbot?count={}&width={}&height={}",
+                count, width, height
+            ))?
+            .json()?)
+        } else {
+            Err(Error::WidthHeightOutOfRange)
+        }
     } else {
         Err(Error::CountOutOfRange)
     }
@@ -107,13 +212,18 @@ pub enum Error {
     Reqwest(reqwest::Error),
     /// Occurs when the count is higher then 1000 or lower then 1.
     CountOutOfRange,
+    /// Occurs when width and/or height is higher then 100,000 or lower then 10.
+    WidthHeightOutOfRange,
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match *self {
+        match self {
             Error::Reqwest(ref err) => err.fmt(f),
             Error::CountOutOfRange => write!(f, "count must be between 1 and 1000"),
+            Error::WidthHeightOutOfRange => {
+                write!(f, "width and height must be between 10 and 100,000")
+            }
         }
     }
 }
@@ -122,16 +232,17 @@ impl error::Error for Error {
     // Usage of description is soft-deprecated; use Display
     #[cfg(ErrorDescription)]
     fn description(&self) -> &str {
-        match *self {
+        match self {
             Error::Reqwest(ref err) => err.description(),
             Error::CountOutOfRange => "count must be between 1 and 1000",
+            Error::WidthHeightOutOfRange => "width and height must be between 10 and 100,000",
         }
     }
 
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match *self {
+        match self {
             Error::Reqwest(ref err) => err.source(),
-            Error::CountOutOfRange => None,
+            Error::CountOutOfRange | Error::WidthHeightOutOfRange => None,
         }
     }
 }
@@ -154,9 +265,11 @@ mod tests {
             colors: vec![
                 Dot {
                     color: Color::new(1.0, 1.0, 1.0, 1.0),
+                    coordinates: None,
                 },
                 Dot {
                     color: Color::new(0.0, 0.0, 0.0, 1.0),
+                    coordinates: None,
                 },
             ],
         };
@@ -171,27 +284,66 @@ mod tests {
     }
 
     #[test]
-    fn test_fetch_count_out_of_range() {
-        assert!(match fetch(-1) {
-            Err(Error::CountOutOfRange) => true,
-            _ => false,
-        });
+    fn test_coordinates() {
+        let hexbot = Hexbot {
+            colors: vec![
+                Dot {
+                    color: Color::new(1.0, 1.0, 1.0, 1.0),
+                    coordinates: Some(Coordinate { x: 100, y: 100 }),
+                },
+                Dot {
+                    color: Color::new(1.0, 1.0, 1.0, 1.0),
+                    coordinates: Some(Coordinate { x: 500, y: 30 }),
+                },
+            ],
+        };
+        assert_eq!(hexbot.coordinates(), Some(vec![(100, 100), (500, 30)]));
+
+        let hexbot = Hexbot {
+            colors: vec![
+                Dot {
+                    color: Color::new(1.0, 1.0, 1.0, 1.0),
+                    coordinates: None,
+                },
+                Dot {
+                    color: Color::new(0.0, 0.0, 0.0, 1.0),
+                    coordinates: None,
+                },
+            ],
+        };
+        assert_eq!(hexbot.coordinates(), None);
+    }
+
+    #[test]
+    fn test_fetch() {
+        // check that 0 results in Error::CountOutOfRange
         assert!(match fetch(0) {
             Err(Error::CountOutOfRange) => true,
             _ => false,
         });
+        // check that 1001 results in Error::CountOutOfRange
         assert!(match fetch(1001) {
             Err(Error::CountOutOfRange) => true,
             _ => false,
         });
-    }
-
-    #[test]
-    #[ignore]
-    fn test_fetch() {
-        let mut hb = fetch(3).unwrap();
-        assert_eq!(hb.colors.len(), 3);
-        hb.colors = vec![];
-        assert_eq!(hb, Hexbot { colors: vec![] });
+        // check that 1 does not result in Error::CountOutOfRange
+        assert!(match fetch(1) {
+            Err(Error::CountOutOfRange) => false,
+            _ => true,
+        });
+        // check that 1000 does not result in Error::CountOutOfRange
+        assert!(match fetch(1000) {
+            Err(Error::CountOutOfRange) => false,
+            _ => true,
+        });
+        // Check that the Ok() data type of fetch is hexbot.
+        // Note that this is not done during execution, it is done during compilation.
+        #[allow(unused_variables)]
+        match fetch(3) {
+            Ok(hb) => {
+                let hexbot: Hexbot = hb;
+            }
+            Err(err) => (),
+        }
     }
 }