+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6]
+### Added
+ * `Hexbot.clone()` (derive trait).
+ * support for the width and height parameters of hexbot.
+   * `Hexbot.coordinates()`
+   * `Hexbot.has_coordinates()`
+   * `request_api::fetch_with_coordinates()`
+   * `Error::WidthHeightOutOfRange`
+
+### Changed
+ * `fmt::Display` for `Hexbot`.
+ * Improved docs & Hacking.
+
 ## [0.0.5]
 ### Added
  * custom error type.
@@ -47,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Support for requesting, parsing and printing a hexbot request without parameters.
  - All stuff around a project like README, LICENSE, .gitignore, ...
 
+[0.0.6]: https://github.com/rusty-snake/hexbot/tree/v0.0.6
 [0.0.5]: https://github.com/rusty-snake/hexbot/tree/v0.0.5
 [0.0.4]: https://github.com/rusty-snake/hexbot/tree/v0.0.4
 [0.0.3]: https://github.com/rusty-snake/hexbot/tree/v0.0.3